# backend/internal/database/dbgen/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package dbgen

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type BackgroundJob struct {
	ID             int64              `json:"id"`
	JobType        string             `json:"job_type"`
	Payload        []byte             `json:"payload"`
	IdempotencyKey pgtype.Text        `json:"idempotency_key"`
	State          string             `json:"state"`
	Attempts       int32              `json:"attempts"`
	MaxAttempts    int32              `json:"max_attempts"`
	RunAt          pgtype.Timestamptz `json:"run_at"`
	LockedAt       pgtype.Timestamptz `json:"locked_at"`
	LockedBy       pgtype.Text        `json:"locked_by"`
	LastError      pgtype.Text        `json:"last_error"`
	CreatedAt      pgtype.Timestamptz `json:"created_at"`
	UpdatedAt      pgtype.Timestamptz `json:"updated_at"`
}

type Device struct {
	ID               pgtype.UUID        `json:"id"`
	UserID           pgtype.UUID        `json:"user_id"`
	InstallationID   string             `json:"installation_id"`
	Platform         string             `json:"platform"`
	PushSubscription []byte             `json:"push_subscription"`
	LastSeenAt       pgtype.Timestamptz `json:"last_seen_at"`
	CreatedAt        pgtype.Timestamptz `json:"created_at"`
	UpdatedAt        pgtype.Timestamptz `json:"updated_at"`
}

type ItineraryShare struct {
	ID        pgtype.UUID        `json:"id"`
	JourneyID pgtype.UUID        `json:"journey_id"`
	OwnerID   pgtype.UUID        `json:"owner_id"`
	TokenHash []byte             `json:"token_hash"`
	Snapshot  []byte             `json:"snapshot"`
	ExpiresAt pgtype.Timestamptz `json:"expires_at"`
	RevokedAt pgtype.Timestamptz `json:"revoked_at"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
}

type Journey struct {
	ID              pgtype.UUID        `json:"id"`
	OwnerID         pgtype.UUID        `json:"owner_id"`
	ClientRequestID pgtype.UUID        `json:"client_request_id"`
	Kind            string             `json:"kind"`
	Label           string             `json:"label"`
	Status          string             `json:"status"`
	StartedAt       pgtype.Timestamptz `json:"started_at"`
	EndedAt         pgtype.Timestamptz `json:"ended_at"`
	Version         int32              `json:"version"`
	CreatedAt       pgtype.Timestamptz `json:"created_at"`
	UpdatedAt       pgtype.Timestamptz `json:"updated_at"`
}

type JourneyStop struct {
	ID              pgtype.UUID        `json:"id"`
	JourneyID       pgtype.UUID        `json:"journey_id"`
	PlaceID         pgtype.UUID        `json:"place_id"`
	SequenceNumber  int64              `json:"sequence_number"`
	CapturedAt      pgtype.Timestamptz `json:"captured_at"`
	DisplayName     string             `json:"display_name"`
	Note            string             `json:"note"`
	Location        interface{}        `json:"location"`
	CreatedAt       pgtype.Timestamptz `json:"created_at"`
	UpdatedAt       pgtype.Timestamptz `json:"updated_at"`
	ClientRequestID pgtype.UUID        `json:"client_request_id"`
}

type LocationSample struct {
	ID              int64              `json:"id"`
	JourneyID       pgtype.UUID        `json:"journey_id"`
	UserID          pgtype.UUID        `json:"user_id"`
	SampleID        pgtype.UUID        `json:"sample_id"`
	CapturedAt      pgtype.Timestamptz `json:"captured_at"`
	Location        interface{}        `json:"location"`
	AccuracyM       pgtype.Float4      `json:"accuracy_m"`
	SpeedMps        pgtype.Float4      `json:"speed_mps"`
	HeadingDeg      pgtype.Float4      `json:"heading_deg"`
	IsAccepted      bool               `json:"is_accepted"`
	RejectionReason pgtype.Text        `json:"rejection_reason"`
	CreatedAt       pgtype.Timestamptz `json:"created_at"`
}

type OutboxEvent struct {
	ID            int64              `json:"id"`
	AggregateType string             `json:"aggregate_type"`
	AggregateID   pgtype.UUID        `json:"aggregate_id"`
	EventType     string             `json:"event_type"`
	Payload       []byte             `json:"payload"`
	CreatedAt     pgtype.Timestamptz `json:"created_at"`
	PublishedAt   pgtype.Timestamptz `json:"published_at"`
}

type Photo struct {
	ID              pgtype.UUID        `json:"id"`
	OwnerID         pgtype.UUID        `json:"owner_id"`
	JourneyID       pgtype.UUID        `json:"journey_id"`
	StopID          pgtype.UUID        `json:"stop_id"`
	ObjectKey       string             `json:"object_key"`
	ThumbnailKey    pgtype.Text        `json:"thumbnail_key"`
	ContentType     string             `json:"content_type"`
	ByteSize        pgtype.Int8        `json:"byte_size"`
	Status          string             `json:"status"`
	CapturedAt      pgtype.Timestamptz `json:"captured_at"`
	CreatedAt       pgtype.Timestamptz `json:"created_at"`
	UpdatedAt       pgtype.Timestamptz `json:"updated_at"`
	ClientRequestID pgtype.UUID        `json:"client_request_id"`
	ChecksumSha256  pgtype.Text        `json:"checksum_sha256"`
}

type Place struct {
	ID              pgtype.UUID        `json:"id"`
	OwnerID         pgtype.UUID        `json:"owner_id"`
	Name            string             `json:"name"`
	Notes           string             `json:"notes"`
	Location        interface{}        `json:"location"`
	CreatedAt       pgtype.Timestamptz `json:"created_at"`
	UpdatedAt       pgtype.Timestamptz `json:"updated_at"`
	DeletedAt       pgtype.Timestamptz `json:"deleted_at"`
	ClientRequestID pgtype.UUID        `json:"client_request_id"`
}

type ReverseGeocodeCache struct {
	LatitudeE5      int32              `json:"latitude_e5"`
	LongitudeE5     int32              `json:"longitude_e5"`
	DisplayName     string             `json:"display_name"`
	ProviderPayload []byte             `json:"provider_payload"`
	CreatedAt       pgtype.Timestamptz `json:"created_at"`
	UpdatedAt       pgtype.Timestamptz `json:"updated_at"`
}

type RouteSegment struct {
	ID            pgtype.UUID        `json:"id"`
	JourneyID     pgtype.UUID        `json:"journey_id"`
	SegmentNumber int32              `json:"segment_number"`
	StartedAt     pgtype.Timestamptz `json:"started_at"`
	EndedAt       pgtype.Timestamptz `json:"ended_at"`
	RawPath       interface{}        `json:"raw_path"`
	MatchedPath   interface{}        `json:"matched_path"`
	DistanceM     float64            `json:"distance_m"`
	DurationS     int32              `json:"duration_s"`
	CreatedAt     pgtype.Timestamptz `json:"created_at"`
	UpdatedAt     pgtype.Timestamptz `json:"updated_at"`
}

type User struct {
	ID                   pgtype.UUID        `json:"id"`
	OidcSubject          string             `json:"oidc_subject"`
	DisplayName          string             `json:"display_name"`
	DefaultTransportMode string             `json:"default_transport_mode"`
	CreatedAt            pgtype.Timestamptz `json:"created_at"`
	UpdatedAt            pgtype.Timestamptz `json:"updated_at"`
}